+++ src/anchors/PerceptorTabAnchor.tsx
@@ -1,14 +1,14 @@
 import $ from 'jquery';
 import { utils } from 'github-url-detection';
 import { isPerceptor, runsWhen, isPublicRepo } from '../utils/utils';
-import PerceptorBase from './PerceptorBase';
+import PerceptorBase from '../PerceptorBase';
 import { render } from 'react-dom';
 import React from 'react';
-import TeachingBubbleWrapper from '../views/TeachingBubbleWrapper';
+import TeachingBubbleWrapper from '../views/TeachingBubbleWrapperView/TeachingBubbleWrapperView';
 import logger from '../utils/logger';
 
 @runsWhen([isPublicRepo])
-class PerceptorTab extends PerceptorBase {
+class PerceptorTabAnchor extends PerceptorBase {
   public async run(): Promise<void> {
     // avoid redundant clone
     let perceptorTab = $('#perceptor_tab');
@@ -95,4 +95,4 @@ class PerceptorTab extends PerceptorBase {
   }
 }
 
-export default PerceptorTab;
+export default PerceptorTabAnchor;
