+++ src/pages/ContentScripts/index.ts
@@ -13,30 +13,29 @@ injected.onload = function () {
 
 // initializeIcons() should only be called once per app and must be called before rendering any components.
 import { initializeIcons } from '@fluentui/react/lib/Icons';
-import DeveloperActiInflTrend from '../../anchors/DeveloperActInflTrend';
-import RepoActiInflTrend from '../../anchors/RepoActiInflTrend';
-import PerceptorTab from '../../anchors/PerceptorTab';
-import PerceptorLayout from '../../anchors/PerceptorLayout';
-import DeveloperNetwork from '../../anchors/DeveloperNetwork';
-import ProjectNetwork from '../../anchors/ProjectNetwork';
-import Hypertrons from '../../anchors/Hypertrons';
-import { inject2Perceptor, Perceptor } from '../../anchors/Perceptor';
 initializeIcons();
-inject2Perceptor(DeveloperActiInflTrend);
-inject2Perceptor(RepoActiInflTrend);
+
+import DeveloperActiInflTrendAnchor from '../../anchors/DeveloperActInflTrendAnchor';
+import RepoActiInflTrendAnchor from '../../anchors/RepoActiInflTrendAnchor';
+import PerceptorTabAnchor from '../../anchors/PerceptorTabAnchor';
+import PerceptorLayoutAnchor from '../../anchors/PerceptorLayoutAnchor';
+import DeveloperNetworkAnchor from '../../anchors/DeveloperNetworkAnchor';
+import ProjectNetworkAnchor from '../../anchors/ProjectNetworkAnchor';
+import HypertronsAnchor from '../../anchors/HypertronsAnchor';
+import { inject2Perceptor, Perceptor } from '../../Perceptor';
 
 import { loadSettings } from '../../utils/settings';
 
 import './index.css';
 
 // inject to Perceptor's static variable
-inject2Perceptor(DeveloperActiInflTrend);
-inject2Perceptor(RepoActiInflTrend);
-inject2Perceptor(PerceptorTab);
-inject2Perceptor(PerceptorLayout);
-inject2Perceptor(DeveloperNetwork);
-inject2Perceptor(ProjectNetwork);
-inject2Perceptor(Hypertrons);
+inject2Perceptor(DeveloperActiInflTrendAnchor);
+inject2Perceptor(RepoActiInflTrendAnchor);
+inject2Perceptor(PerceptorTabAnchor);
+inject2Perceptor(PerceptorLayoutAnchor);
+inject2Perceptor(DeveloperNetworkAnchor);
+inject2Perceptor(ProjectNetworkAnchor);
+inject2Perceptor(HypertronsAnchor);
 
 async function mainInject() {
   const settings = await loadSettings();
